Create rtu-test-server.js
In this updated script, we use the modbus-serial library to communicate with the slave devices using the Modbus RTU protocol over TCP. We first connect to the MOXA gateway using the connectTCP method and set the serial communication parameters using the setRTUConfig method. We then periodically read the required data for all USS addresses and store it in the data object.

We expose the data object through a REST endpoint /TR using Express.js. The server listens on port 3000.